Move fundraiser redux helpers to the reducer file
